Add delay between volume changes
To repeat volume commands a small delay is needed between the commands.
A delay of at least 0.1 seconds seems to be needed to get the repeated
command accepted. Longer delays also work, but cause long command
sequences.

For STR-DB2000 there seems to be a "turbo" mode. If the delay is between
0.05 seconds and 0.08 seconds, then the volume changes quickly. But the
volume change depends on the current volume. The steps seem to be:

...
-34 dB
-36 dB
-38 dB - 2 dB steps for higher volumes
-43 dB - 5 dB steps for lower volumes
-48 dB
-53 dB
...

(I'm not trying to use the turbo mode yet, but could possibly be used to
implement setting specific volumes)